fix: update visualizations
